Server: Extract raw api to be implemented in Project1 into raw_api.go (#285)
* Extract raw api to be implemented in Project1 into raw_api.go

* Run go fmt to apply automatic formatter and delete redundant comments

* Add comments and hints to each function
